Added friends, messages, and user data
I have added csv files that contain real data for friends, messages, and
users for our previous version. I have also pushed a filed called
"schema" which lists the schema for each table and also explains the
idea of a "foreign key/id". Here is a wiki on what that is if you have
never done relational database with work foreign keys:
https://en.wikipedia.org/wiki/Foreign_key