Start on making the sliders actually work
I messed around with a whole bunch of stuff, including making a new
Component that will hold the dropdown and the slider and starting to
hook that up. That was mostly it, actually. Also I hooked up the
TweetFilterer object properly. I still need to do some stuff to make it
actually work but it's getting closer.