Add caching
Basically I refactored `Tweet`s to be different from the thing that
renders a `Tweet` (a `TweetView`). `Tweet`s hold all the info about a
tweet and it's all calculated on startup. It's way faster now. It's
still slow sometimes but it's a huge improvement. Why did it take me a
month to do this? The mystery indeed...